chore: rename browsers folders and files
src/browsers to src/browser-data to avoid confusion
between the top-level package and inner folder. Also, changed
imports to always use browser-data entrypoint for imports.

browsers.ts to main-cli.ts to indicate that the file is an entrypoint
to the CLI.
