+++ packages/browsers/package.json
@@ -9,7 +9,7 @@
     "test": "wireit"
   },
   "bin": {
-    "@puppeteer/browsers": "lib/cjs/browsers.js"
+    "@puppeteer/browsers": "lib/cjs/main-cli.js"
   },
   "main": "./lib/cjs/main.js",
   "exports": {
@@ -20,7 +20,7 @@
   },
   "wireit": {
     "build": {
-      "command": "tsc -b && chmod a+x lib/cjs/browsers.js lib/esm/browsers.js",
+      "command": "tsc -b && chmod a+x lib/cjs/main-cli.js lib/esm/main-cli.js",
       "files": [
         "src/**/*.ts",
         "tsconfig.json"
