+++ packages/browsers/src/browser-data/firefox.ts
