+++ packages/browsers/src/browser-data/types.ts
