+++ packages/browsers/test/src/firefox-data.spec.ts
@@ -21,12 +21,12 @@ import path from 'path';
 
 import rimraf from 'rimraf';
 
-import {BrowserPlatform} from '../../lib/cjs/browsers/browsers.js';
+import {BrowserPlatform} from '../../lib/cjs/browser-data/browser-data.js';
 import {
   createProfile,
   relativeExecutablePath,
   resolveDownloadUrl,
-} from '../../lib/cjs/browsers/firefox.js';
+} from '../../lib/cjs/browser-data/firefox.js';
 
 describe('Firefox', () => {
   it('should resolve download URLs', () => {
